+++ lib/src/dummy_logger.dart
@@ -5,13 +5,5 @@ import 'logger.dart';
 /// Useful in tests or environments where logging should be silenced entirely.
 class DummyLogger extends Logger {
   @override
-  void write(
-    LogLevel level,
-    String className,
-    String message, [
-    Object? error,
-    StackTrace? stackTrace,
-  ]) {
-    // Do nothing
-  }
+  void write(LogEntry entry) {}
 }