+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.1.0
 - Added `notice` log level between `info` and `warn`.
+- Added `log.obs.event` for logging structured observability events.
+- Added `log.obs.metric` for logging numeric observability metrics.
 
 ## 1.0.0
 - Initial version.